prefect_pipeline/requirements.txt
prefect>=3.2.0
mlflow
tensorflow
sqlalchemy
psycopg2-binary
pillow
numpy
